refactor: move tag to ui
